코드 수정

+++ src/main/java/com/soloproject/shoppingmall/cart/dto/CartResponseDto.java
@@ -13,6 +13,6 @@ import java.util.List;
 public class CartResponseDto extends Auditable {
     private long cartId;
     private long memberId;
-    private long totalPrice;
+    private int totalPrice;
     private List<CartProductResponseDto> cartProducts;
 }
