+++ src/main/java/com/soloproject/shoppingmall/cart/entity/Cart.java
@@ -2,7 +2,6 @@ package com.soloproject.shoppingmall.cart.entity;
 
 import com.soloproject.shoppingmall.audit.Auditable;
 import com.soloproject.shoppingmall.member.entity.Member;
-import com.soloproject.shoppingmall.order.entity.Order;
 import jakarta.persistence.*;
 import lombok.AllArgsConstructor;
 import lombok.Getter;
@@ -24,7 +23,7 @@ public class Cart extends Auditable {
     private long cartId;
 
     @Column(nullable = false)
-    private long totalPrice;
+    private int totalPrice;
 
     @OneToOne
     @JoinColumn(name = "member_id")
